3/dd2420_mrf_tutorial3_220110/dd2420_mrf_tutorial3/requirements.txt

matplotlib==3.5.0
numpy==1.21.2 
Pillow==8.4.0
